Add Supplier<String> based logger methods
This means the caller doesn't need to check if the logger is enabled prior to doing
a long running log message calculation.  The calculation is only executed if the level
is high enough to be logged.

Feel free to ignore this if it adds unnecessary complication to the otherwise simple api